wip: welp it compiles
