+++ tests/Unleash.Tests/Mock/MockFileSystem.cs
@@ -1,10 +1,13 @@
 ﻿using System.IO;
+using System.Text;
 using Unleash.Internal;
 
 namespace Unleash.Tests.Mock
 {
     class MockFileSystem : IFileSystem
     {
+        public Encoding Encoding => throw new NotImplementedException();
+
         public bool FileExists(string path)
         {
             return true;
@@ -27,6 +30,18 @@ namespace Unleash.Tests.Mock
         public string ReadAllText(string path)
         {
             return string.Empty;
+        }
+
+        public void Move(string sourcePath, string destPath)
+        {
+        }
+
+        public void Replace(string sourceFileName, string destinationFileName, string destinationBackupFileName)
+        {
+        }
+
+        public void Delete(string path)
+        {
         }
     }
 }
